performed my first code refactoring like a very tiny development cicle
basically realised that it would be better if every StockReader handled only the part of processing the data source (a .csv file, a .json file, a remote web service, etc.) and the rest of information processing (extracting title, dates, facts, etc.) should be handled by a whole StockData hierarchy. After that, a Dependency Injection like design could be applied.